ed:wq

+++ 7thfeb/lab20Task.py
@@ -1,5 +1,5 @@
 import math
-a = int(input("enter a number\n"))
+a = float(input("enter a number\n"))
 sqr = math.sqrt(a)
 cube = math.pow(a, 1/3)
 print("squre of entered number is:" ,sqr)
